modularized the ProfileAnalytics Component
